chore: created the anakin and sebulba folders
